Sprint 22. 9-gateway-microservices

+++ commerce/warehouse/src/main/java/ru/yandex/practicum/warehouse/model/WarehouseProduct.java
@@ -25,7 +25,7 @@ public class WarehouseProduct {
     @NotNull
     @Column(nullable = false)
     @Builder.Default
-    Integer quantity = 0;
+    Long quantity = 0L;
 
     @NotNull
     @Positive
